Add initial AppKitExtensions local package (#3082)
Task/Issue URL:
https://app.asana.com/0/1201037661562251/1207999635472099/f
Tech Design URL:
CC:

**Description**:
As title.
The most immediate purpose is so it can be used in DBP. It also seems we
were doing some crimes to include the NSApplication extension in the net
p targets, so this PR also unwinds that.
I've made notes inline where I think specific scrutiny is warranted.

**Steps to test this PR**:
1. Smoke test the app and specifically the VPN to ensure nothing breaks

<!--
Tagging instructions
If this PR isn't ready to be merged for whatever reason it should be
marked with the `DO NOT MERGE` label (particularly if it's a draft)
If it's pending Product Review/PFR, please add the `Pending Product
Review` label.

If at any point it isn't actively being worked on/ready for
review/otherwise moving forward (besides the above PR/PFR exception)
strongly consider closing it (or not opening it in the first place). If
you decide not to close it, make sure it's labelled to make it clear the
PRs state and comment with more information.
-->

**Definition of Done**:

* [ ] Does this PR satisfy our [Definition of
Done](https://app.asana.com/0/1202500774821704/1207634633537039/f)?

---
###### Internal references:
[Pull Request Review
Checklist](https://app.asana.com/0/1202500774821704/1203764234894239/f)
[Software Engineering
Expectations](https://app.asana.com/0/59792373528535/199064865822552)
[Technical Design
Template](https://app.asana.com/0/59792373528535/184709971311943)
[Pull Request
Documentation](https://app.asana.com/0/1202500774821704/1204012835277482/f)

---------
