+++ DuckDuckGoVPN/DuckDuckGoVPNAppDelegate.swift
@@ -440,17 +440,3 @@ final class DuckDuckGoVPNAppDelegate: NSObject, NSApplicationDelegate {
         }
     }
 }
-
-extension NSApplication {
-
-    enum RunType: Int, CustomStringConvertible {
-        case normal
-        var description: String {
-            switch self {
-            case .normal: return "normal"
-            }
-        }
-    }
-    static var runType: RunType { .normal }
-
-}
