+++ LocalPackages/AppKitExtensions/README.md
@@ -1,0 +1,6 @@
+# AppKitExtensions
+
+A package for any extensions on AppKit classes, so local packages and other targets besides the main app can use them.
+  
+- If you need to add a new one put it here unless it's specific to the main app.  
+- If you need to modify on existing one, consider moving it here.
